+++ weeklyMission1/member.js
@@ -1,0 +1,2 @@
+export const MEMBER_ID = "test@codeit.com";
+export const MEMBER_PASSWORD = "codeit101";
