+++ weeklyMission1/signup.html
@@ -27,20 +27,20 @@
         <div class="signin-password">
             <p>비밀번호</p>
             <input class="signup-password-input" id="signInPw" type="password" placeholder="password">
-            <img src="./img/unhidden.svg" class="password-img" alt="opened-eye">
+            <img src="./img/hidden.svg" class="password-img" alt="closed-eye">
             <div class="signup-password-error-msg"></div>
         </div>
         <div class="signin-password">
             <p>비밀번호확인</p>
             <input class="signup-password-check-input" id="signInPw" type="password" placeholder="password">
-            <img src="./img/unhidden.svg" class="password-img-check" alt="opened-eye-second">
+            <img src="./img/hidden.svg" class="password-img-check" alt="closed-eye-second">
             <div class="password-check-error-msg"></div>
         </div>
-        <a href="/">
-            <div class="signin-login">
+        <!-- <a href="/"> -->
+            <div class="signup-btn">
                 회원가입
             </div>
-        </a>
+        <!-- </a> -->
         <div class="signin-social-login">
             <p>다른 방식으로 가입하기</p>
             <div class="signin-social-login-icons">
