+++ weeklyMission1/tags.js
@@ -1,0 +1,20 @@
+const signUpEmailInput = document.querySelector('.signup-email-input');
+const signUpEmailErrorMsg = document.querySelector('.signup-email-error-msg');
+const signUpPasswordInput = document.querySelector('.signup-password-input');
+const signUpPasswordCheckInput = document.querySelector('.signup-password-check-input');
+const signUpPasswordErrorMsg = document.querySelector('.signup-password-error-msg');
+const passwordCheckErrorMsg = document.querySelector('.password-check-error-msg');
+const passwordImg = document.querySelector('.password-img');
+const passwordImgCheck = document.querySelector('.password-img-check');
+const signUpBtn = document.querySelector('.signup-btn');
+
+const signInEmailInput = document.querySelector('.signin-email-input');
+const signInPasswordInput = document.querySelector('.signin-password-input');
+const emailErrorMsg = document.querySelector('.email-error-msg');
+const passwordErrorMsg = document.querySelector('.password-error-msg');
+const signInLogin = document.querySelector('.signin-login');
+
+export { signUpEmailInput, signUpEmailErrorMsg, signUpPasswordInput, signUpPasswordCheckInput, signUpPasswordErrorMsg, 
+        passwordCheckErrorMsg, passwordImg, passwordImgCheck, signUpBtn };
+
+export { signInEmailInput, signInPasswordInput, emailErrorMsg, passwordErrorMsg, signInLogin };
